Add preference to set the pkg-config binary path
There was a confusion in the code between the PKG_CONFIG_PATH
variable and the path to the pkg-config binary. The former was used
as the location for the pkg-config binary which is wrong.

This commit creates an additional preference to set the location of
the pkg-config binary and uses it to call the pkg-config binary.

Signed-off-by: Mélanie Bats <melanie.bats@obeo.fr>